refactor: 패키지 분리 방식 변경
