+++ src/main/java/roomescape/reservation/presentation/controller/ThemeController.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.presentation;
+package roomescape.reservation.presentation.controller;
 
 import java.net.URI;
 import java.util.List;
